third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/block-does-not-leak-referrer-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor blocked access to 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?enable-full-block=1&q=%3Cscript%3Ealert(String.fromCharCode(0x58,0x53,0x53))%3C/script%3E' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
PASS frame.contentDocument is null
PASS successfullyParsed is true

TEST COMPLETE
